requirements.txt
aiogram>=3.0.0
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
redis
aiohttp
openai
huggingface_hub
python-dotenv==1.0.1
pydantic==2.9.2
orjson==3.10.12
mediapipe
opencv-python-headless
numpy
